AlertMessage now a separte new component
